+++ core/core-api/src/test/groovyss/com/hbc/pms/core/api/controllers/AlarmConditionControllerQueryEndpointsFunctionalSpec.groovy
