+++ core/core-api/src/test/groovyss/com/hbc/pms/core/api/services/AlarmResolverFunctionalSpec.groovy
